Move RequestOptions and Util into Util namespace
