+++ tests/BalanceTest.php
@@ -9,7 +9,7 @@ class BalanceTest extends TestCase
         self::authorizeFromEnv();
         $d = Balance::retrieve();
         $this->assertSame($d->object, "balance");
-        $this->assertTrue(Util::isList($d->available));
-        $this->assertTrue(Util::isList($d->pending));
+        $this->assertTrue(Util\Util::isList($d->available));
+        $this->assertTrue(Util\Util::isList($d->pending));
     }
 }
